# gen/http/analyzer/server/encode_decode.go
// Code generated by goa, DO NOT EDIT.
//
// analyzer HTTP server encoders and decoders
//
// Command:
// $ goa gen github.com/hatyibei/shingan/design

package server

import (
	"context"
	"errors"
	"io"
	"net/http"

	analyzer "github.com/hatyibei/shingan/gen/analyzer"
	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeAnalyzeResponse returns an encoder for responses returned by the
// analyzer analyze endpoint.
func EncodeAnalyzeResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*analyzer.AnalysisResult)
		enc := encoder(ctx, w)
		body := NewAnalyzeResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeAnalyzeRequest returns a decoder for requests sent to the analyzer
// analyze endpoint.
func DecodeAnalyzeRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*analyzer.AnalyzePayload, error) {
	return func(r *http.Request) (*analyzer.AnalyzePayload, error) {
		var payload *analyzer.AnalyzePayload
		var (
			body AnalyzeRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if errors.Is(err, io.EOF) {
				return payload, goa.MissingPayloadError()
			}
			var gerr *goa.ServiceError
			if errors.As(err, &gerr) {
				return payload, gerr
			}
			return payload, goa.DecodePayloadError(err.Error())
		}
		err = ValidateAnalyzeRequestBody(&body)
		if err != nil {
			return payload, err
		}
		payload = NewAnalyzePayload(&body)

		return payload, nil
	}
}

// EncodeAnalyzeError returns an encoder for errors returned by the analyze
// analyzer endpoint.
func EncodeAnalyzeError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "invalid_format":
			var res *goa.ServiceError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewAnalyzeInvalidFormatResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusBadRequest)
			return enc.Encode(body)
		case "parse_error":
			var res *goa.ServiceError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewAnalyzeParseErrorResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusUnprocessableEntity)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// EncodeHealthResponse returns an encoder for responses returned by the
// analyzer health endpoint.
func EncodeHealthResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(string)
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// marshalAnalyzerFindingToFindingResponseBody builds a value of type
// *FindingResponseBody from a value of type *analyzer.Finding.
func marshalAnalyzerFindingToFindingResponseBody(v *analyzer.Finding) *FindingResponseBody {
	res := &FindingResponseBody{
		Rule:       v.Rule,
		Severity:   v.Severity,
		NodeID:     v.NodeID,
		Message:    v.Message,
		Suggestion: v.Suggestion,
	}

	return res
}

// marshalAnalyzerSummaryToSummaryResponseBody builds a value of type
// *SummaryResponseBody from a value of type *analyzer.Summary.
func marshalAnalyzerSummaryToSummaryResponseBody(v *analyzer.Summary) *SummaryResponseBody {
	res := &SummaryResponseBody{
		Total:    v.Total,
		Critical: v.Critical,
		Warning:  v.Warning,
		Info:     v.Info,
	}

	return res
}