# gen/http/analyzer/server/types.go
// Code generated by goa, DO NOT EDIT.
//
// analyzer HTTP server types
//
// Command:
// $ goa gen github.com/hatyibei/shingan/design

package server

import (
	analyzer "github.com/hatyibei/shingan/gen/analyzer"
	goa "goa.design/goa/v3/pkg"
)

// AnalyzeRequestBody is the type of the "analyzer" service "analyze" endpoint
// HTTP request body.
type AnalyzeRequestBody struct {
	// Input format: json or adk-go
	Format *string `form:"format,omitempty" json:"format,omitempty" xml:"format,omitempty"`
	// Raw workflow definition content
	Content []byte `form:"content,omitempty" json:"content,omitempty" xml:"content,omitempty"`
	// Output format: json or markdown
	Output *string `form:"output,omitempty" json:"output,omitempty" xml:"output,omitempty"`
}

// AnalyzeResponseBody is the type of the "analyzer" service "analyze" endpoint
// HTTP response body.
type AnalyzeResponseBody struct {
	// Detected issues
	Findings []*FindingResponseBody `form:"findings" json:"findings" xml:"findings"`
	// Severity counts
	Summary *SummaryResponseBody `form:"summary" json:"summary" xml:"summary"`
	// Exit code analog: 0/1/2
	ExitCode int `form:"exit_code" json:"exit_code" xml:"exit_code"`
}

// AnalyzeInvalidFormatResponseBody is the type of the "analyzer" service
// "analyze" endpoint HTTP response body for the "invalid_format" error.
type AnalyzeInvalidFormatResponseBody struct {
	// Name is the name of this class of errors.
	Name string `form:"name" json:"name" xml:"name"`
	// ID is a unique identifier for this particular occurrence of the problem.
	ID string `form:"id" json:"id" xml:"id"`
	// Message is a human-readable explanation specific to this occurrence of the
	// problem.
	Message string `form:"message" json:"message" xml:"message"`
	// Is the error temporary?
	Temporary bool `form:"temporary" json:"temporary" xml:"temporary"`
	// Is the error a timeout?
	Timeout bool `form:"timeout" json:"timeout" xml:"timeout"`
	// Is the error a server-side fault?
	Fault bool `form:"fault" json:"fault" xml:"fault"`
}

// AnalyzeParseErrorResponseBody is the type of the "analyzer" service
// "analyze" endpoint HTTP response body for the "parse_error" error.
type AnalyzeParseErrorResponseBody struct {
	// Name is the name of this class of errors.
	Name string `form:"name" json:"name" xml:"name"`
	// ID is a unique identifier for this particular occurrence of the problem.
	ID string `form:"id" json:"id" xml:"id"`
	// Message is a human-readable explanation specific to this occurrence of the
	// problem.
	Message string `form:"message" json:"message" xml:"message"`
	// Is the error temporary?
	Temporary bool `form:"temporary" json:"temporary" xml:"temporary"`
	// Is the error a timeout?
	Timeout bool `form:"timeout" json:"timeout" xml:"timeout"`
	// Is the error a server-side fault?
	Fault bool `form:"fault" json:"fault" xml:"fault"`
}

// FindingResponseBody is used to define fields on response body types.
type FindingResponseBody struct {
	Rule       string  `form:"rule" json:"rule" xml:"rule"`
	Severity   string  `form:"severity" json:"severity" xml:"severity"`
	NodeID     string  `form:"node_id" json:"node_id" xml:"node_id"`
	Message    string  `form:"message" json:"message" xml:"message"`
	Suggestion *string `form:"suggestion,omitempty" json:"suggestion,omitempty" xml:"suggestion,omitempty"`
}

// SummaryResponseBody is used to define fields on response body types.
type SummaryResponseBody struct {
	Total    int `form:"total" json:"total" xml:"total"`
	Critical int `form:"critical" json:"critical" xml:"critical"`
	Warning  int `form:"warning" json:"warning" xml:"warning"`
	Info     int `form:"info" json:"info" xml:"info"`
}

// NewAnalyzeResponseBody builds the HTTP response body from the result of the
// "analyze" endpoint of the "analyzer" service.
func NewAnalyzeResponseBody(res *analyzer.AnalysisResult) *AnalyzeResponseBody {
	body := &AnalyzeResponseBody{
		ExitCode: res.ExitCode,
	}
	if res.Findings != nil {
		body.Findings = make([]*FindingResponseBody, len(res.Findings))
		for i, val := range res.Findings {
			if val == nil {
				body.Findings[i] = nil
				continue
			}
			body.Findings[i] = marshalAnalyzerFindingToFindingResponseBody(val)
		}
	} else {
		body.Findings = []*FindingResponseBody{}
	}
	if res.Summary != nil {
		body.Summary = marshalAnalyzerSummaryToSummaryResponseBody(res.Summary)
	}
	return body
}

// NewAnalyzeInvalidFormatResponseBody builds the HTTP response body from the
// result of the "analyze" endpoint of the "analyzer" service.
func NewAnalyzeInvalidFormatResponseBody(res *goa.ServiceError) *AnalyzeInvalidFormatResponseBody {
	body := &AnalyzeInvalidFormatResponseBody{
		Name:      res.Name,
		ID:        res.ID,
		Message:   res.Message,
		Temporary: res.Temporary,
		Timeout:   res.Timeout,
		Fault:     res.Fault,
	}
	return body
}

// NewAnalyzeParseErrorResponseBody builds the HTTP response body from the
// result of the "analyze" endpoint of the "analyzer" service.
func NewAnalyzeParseErrorResponseBody(res *goa.ServiceError) *AnalyzeParseErrorResponseBody {
	body := &AnalyzeParseErrorResponseBody{
		Name:      res.Name,
		ID:        res.ID,
		Message:   res.Message,
		Temporary: res.Temporary,
		Timeout:   res.Timeout,
		Fault:     res.Fault,
	}
	return body
}

// NewAnalyzePayload builds a analyzer service analyze endpoint payload.
func NewAnalyzePayload(body *AnalyzeRequestBody) *analyzer.AnalyzePayload {
	v := &analyzer.AnalyzePayload{
		Format:  *body.Format,
		Content: body.Content,
	}
	if body.Output != nil {
		v.Output = *body.Output
	}
	if body.Output == nil {
		v.Output = "json"
	}

	return v
}

// ValidateAnalyzeRequestBody runs the validations defined on AnalyzeRequestBody
func ValidateAnalyzeRequestBody(body *AnalyzeRequestBody) (err error) {
	if body.Format == nil {
		err = goa.MergeErrors(err, goa.MissingFieldError("format", "body"))
	}
	if body.Content == nil {
		err = goa.MergeErrors(err, goa.MissingFieldError("content", "body"))
	}
	if body.Format != nil {
		if !(*body.Format == "json" || *body.Format == "adk-go") {
			err = goa.MergeErrors(err, goa.InvalidEnumValueError("body.format", *body.Format, []any{"json", "adk-go"}))
		}
	}
	if body.Output != nil {
		if !(*body.Output == "json" || *body.Output == "markdown") {
			err = goa.MergeErrors(err, goa.InvalidEnumValueError("body.output", *body.Output, []any{"json", "markdown"}))
		}
	}
	return
}