# gen/http/openapi3.json
{"openapi":"3.0.3","info":{"title":"Shingan — AI Agent Workflow Static Analyzer API","description":"Static analysis for AI agent workflows","version":"0.1.0"},"servers":[{"url":"http://localhost:8080"}],"paths":{"/analyze":{"post":{"tags":["analyzer"],"summary":"analyze analyzer","operationId":"analyzer#analyze","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnalyzeRequestBody"},"example":{"content":"TW9sZXN0aWFlIGF1dCBvbW5pcyBldCBkb2xvcmUgaWxsdW0gaXN0ZS4=","format":"json","output":"json"}}}},"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnalysisResult"},"example":{"exit_code":5292158723142717433,"findings":[{"message":"Esse aut doloribus.","node_id":"Veritatis neque ea quidem aut qui.","rule":"Voluptatem deleniti dolores rerum.","severity":"critical","suggestion":"Rem perspiciatis."},{"message":"Esse aut doloribus.","node_id":"Veritatis neque ea quidem aut qui.","rule":"Voluptatem deleniti dolores rerum.","severity":"critical","suggestion":"Rem perspiciatis."}],"summary":{"critical":9019864383028453855,"info":7638366186936058184,"total":7539941759466190338,"warning":1596166169809979685}}}}},"400":{"description":"invalid_format: Bad Request response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}},"422":{"description":"parse_error: Unprocessable Entity response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/healthz":{"get":{"tags":["analyzer"],"summary":"health analyzer","operationId":"analyzer#health","responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"string","example":"Doloribus quasi voluptates repellat maxime."},"example":"Aut quo sed at voluptatem sed est."}}}}}}},"components":{"schemas":{"AnalysisResult":{"type":"object","properties":{"exit_code":{"type":"integer","description":"Exit code analog: 0/1/2","example":3685305706602171141,"format":"int64"},"findings":{"type":"array","items":{"$ref":"#/components/schemas/Finding"},"description":"Detected issues","example":[{"message":"Et commodi consequuntur.","node_id":"Quasi voluptatem perferendis.","rule":"Similique in.","severity":"info","suggestion":"Minima accusamus sunt qui ipsum sapiente in."},{"message":"Et commodi consequuntur.","node_id":"Quasi voluptatem perferendis.","rule":"Similique in.","severity":"info","suggestion":"Minima accusamus sunt qui ipsum sapiente in."},{"message":"Et commodi consequuntur.","node_id":"Quasi voluptatem perferendis.","rule":"Similique in.","severity":"info","suggestion":"Minima accusamus sunt qui ipsum sapiente in."},{"message":"Et commodi consequuntur.","node_id":"Quasi voluptatem perferendis.","rule":"Similique in.","severity":"info","suggestion":"Minima accusamus sunt qui ipsum sapiente in."}]},"summary":{"$ref":"#/components/schemas/Summary"}},"example":{"exit_code":5396350171635595078,"findings":[{"message":"Et commodi consequuntur.","node_id":"Quasi voluptatem perferendis.","rule":"Similique in.","severity":"info","suggestion":"Minima accusamus sunt qui ipsum sapiente in."},{"message":"Et commodi consequuntur.","node_id":"Quasi voluptatem perferendis.","rule":"Similique in.","severity":"info","suggestion":"Minima accusamus sunt qui ipsum sapiente in."}],"summary":{"critical":3502243872414327787,"info":6197453941820126429,"total":6517932662113017790,"warning":2146077217038281464}},"required":["findings","summary","exit_code"]},"AnalyzeRequestBody":{"type":"object","properties":{"content":{"type":"string","description":"Raw workflow definition content","example":"Vm9sdXB0YXMgZGlnbmlzc2ltb3MgYWxpYXMgbGliZXJvIHF1aXNxdWFtIGRvbG9yZW1xdWUu","format":"binary"},"format":{"type":"string","description":"Input format: json or adk-go","example":"json","enum":["json","adk-go"]},"output":{"type":"string","description":"Output format: json or markdown","default":"json","example":"markdown","enum":["json","markdown"]}},"example":{"content":"RmFjaWxpcyB2b2x1cHRhcyBwb3JybyBjb25zZXF1YXR1ciBxdW8gbWF4aW1lLg==","format":"json","output":"json"},"required":["format","content"]},"Error":{"type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":false},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"Finding":{"type":"object","properties":{"message":{"type":"string","example":"Fuga ipsam nulla."},"node_id":{"type":"string","example":"Praesentium ex odit labore molestias molestiae."},"rule":{"type":"string","example":"Incidunt nobis minima."},"severity":{"type":"string","example":"warning","enum":["critical","warning","info"]},"suggestion":{"type":"string","example":"Delectus adipisci possimus."}},"example":{"message":"Perspiciatis tempore autem quas aut.","node_id":"Velit voluptatum inventore sint et.","rule":"Sequi recusandae voluptas odit quaerat quisquam qui.","severity":"info","suggestion":"Non quisquam ut est amet temporibus aut."},"required":["rule","severity","node_id","message"]},"Summary":{"type":"object","properties":{"critical":{"type":"integer","example":4513008794304796852,"format":"int64"},"info":{"type":"integer","example":3383490030079515480,"format":"int64"},"total":{"type":"integer","example":6931227284969573035,"format":"int64"},"warning":{"type":"integer","example":2271744659483088426,"format":"int64"}},"example":{"critical":4730167276366075593,"info":6085523593245420613,"total":2253280284776915423,"warning":6737573540959863442},"required":["total","critical","warning","info"]}}},"tags":[{"name":"analyzer","description":"Analyze AI agent workflow definitions for structural issues"}]}